refactor: repository 위치 변경 - domain 하위
