+++ product/src/main/java/com/emotionalcart/product/domain/repository/ReviewImageRepository.java
@@ -1,4 +1,4 @@
-package com.emotionalcart.product.infrastructure;
+package com.emotionalcart.product.domain.repository;
 
 import com.emotionalcart.product.domain.ReviewImage;
 import org.springframework.data.jpa.repository.JpaRepository;
